Add a mechanism for sending messages via meteor
Adds a collection and methods for sending messages to all clients.
These methods are meant to be used when webrtc is not the main part
of the app and is not initialized until needed. In such case, use
these methods to tell the client to turn on the webrtc interface so
it can connect.

 # Please enter the commit message for your changes. Lines starting